Fix NameError on FileNotFoundError
This is a work around to get it working again for other people.
I'll invest more time to working out a better solution when I
have more time to invest. In the mean time, if someone with more
pythong experience with me would like to submit a pull request,
it would be very much appreciated.

Specifically code that was running fine when simply checked out
from the repo and executed, would not work when installed using
`pip`. The error was

    NameError: global name 'FileNotFoundError' is not defined

The package would install without error, but wouild produce that
error when `nt` was executed.